chore: rename rex folder to staking
