+++ src/components/Staking/StakingInfo.vue
@@ -96,14 +96,14 @@ export default defineComponent({
           .col-7 {{ `TOTAL ${symbol} STAKED` }}
           .col-5.text-right.text-weight-bold {{coreRexBalance}}
         .row.q-pt-sm
-          .col-7 STAKED SAVINGS
+          .col-7 SAVINGS
           .col-5.text-right.text-weight-bold {{rexSavings}}
       .col-xs-12.col-sm-6.q-px-lg
         .row
-          .col-7 MATURED STAKED
+          .col-7 MATURED
           .col-5.text-right.text-weight-bold {{maturedRex}}
         .row.q-pt-sm
-          .col-7 MATURING STAKED
+          .col-7 MATURING
           .col-5.text-right.text-weight-bold {{maturingRex}}
 
 </template>
